requirements.txt
google-generativeai==0.7.2
openai==1.40.8